Support ranges starting with a dash again
The support for ranges starting with a dash (e.g. `did -r -3..`) was broken after
an upgrade from optparse to argparse. The reason is that argparse doesn't allow
for option arguments beginning with a dash (https://bugs.python.org/issue9334),
while it used to work well with optparse. To work this around, I added a wrapper
over argparse.ArgumentParser, such that rewrites the arguments so that there's
an equals sign, e.g. turning `-r -3..` into `-r=-3..`.

This fixes issue #25.