Refactor grepping: reduce cross-dependencies
The grepping functionality was spread all over the place: in worklog, session,
interval, as well as in reporting. Now it's scoped only to the new
IntervalFilter class and it's uses in report generation.

While this reduces cross-dependencies and improves overall code quality, this
might potentially cause peformance degradation on super large job log files,
because the regexes are now applied even three times for each interval (check
if a session should be printed, check if interval should be printed, counting
duration of matched intervals), not once as before. If I realize that is a real
problem, I'll find different ways of addressing it.